+++ lib/unit_measurements/unit_groups/all.rb
@@ -2,7 +2,12 @@
 # -*- frozen_string_literal: true -*-
 # -*- warn_indent: true -*-
 
+# Fundamental SI units
+
 require_relative "amount_of_substance"
+require_relative "electric_current"
 require_relative "length"
+require_relative "luminous_intensity"
+require_relative "temperature"
 require_relative "time"
 require_relative "weight"